can update cart

+++ server/controllers/carts.js
@@ -14,22 +14,31 @@ module.exports = {
         let { product_id } = req.params;
         let { id: user_id } = req.session.user;
         req.db.create_cart({ product_id, user_id })
-            .then(() => {
-                req.db.read_cart({ id: user_id })
-                    .then(cart => {
-                        res.status(200).send(cart);
-                    })
-                    .catch(err => {
-                        console.log(err);
-                        res.status(200).send(err);
-                    });
+            .then(cart => {
+                res.status(200).send(cart);
             })
             .catch(err => {
                 console.log(err);
-                res.status(200).send(arr);
+                res.status(200).send(err);
             });
     },
     delete(req, res) {
-        req.db.delete_cart();
+        let { amt } = req.query;
+        let { product_id } = req.params;
+        let { id: user_id } = req.session.user;
+        console.log({ amt, product_id, user_id });
+        let deletion;
+        if (amt)
+            deletion = req.db.delete_cart_amt({ user_id, product_id, amt });
+        else
+            deletion = req.db.delete_cart({ user_id, product_id });
+        deletion
+            .then(cart => {
+                res.status(200).send(cart);
+            })
+            .catch(err => {
+                console.log(err);
+                res.status(200).send(err);
+            });
     }
 }
